机械臂程序

+++ rm/Inc/grip.h
@@ -9,9 +9,9 @@ double up();
 double grip();			//夹住物块
 double loose();			//松开物块
 void rm_servo(double angle);		//云台舵机
-void servo1(double angle);
-void servo2(double angle);
-void servo3(double angle);
+void servo1(float angle);
+void servo2(float angle);
+void servo3(float angle);
 void clip(double angle);			//夹子舵机
